requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiopyston==1.2.1
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
aniso8601==10.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.0.5
asyncio==4.0.0
attrs==25.3.0
autobahn==25.12.2
Automat==25.4.16
babel==2.17.0
banal==1.0.6
bcrypt==4.0.1
beautifulsoup4==4.14.2
billiard==4.2.4
bleach==6.3.0
boto3==1.35.27
botocore==1.35.99
cachelib==0.9.0
cbor2==5.8.0
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmarkgfm==2024.11.20
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
cron_descriptor==2.0.6
cryptography==46.0.3
cycler==0.12.1
daphne==4.0.0
dataset==1.6.2
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Django==4.2
django-axes==5.41.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==25.2
django-ipware==7.0.1
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-ratelimit==4.1.0
django-stubs-ext==5.2.8
django-taggit==6.1.0
PyJWT==2.8.0
django-tasks==0.9.0
django-timezone-field==7.2.1
django-treebeard==4.8.0
djangorestframework==3.14.0
docker==7.0.0
draftjs_exporter==5.1.0
drf-spectacular==0.27.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.124.4
fastjsonschema==2.21.2
ffmpeg==1.4
ffpyplayer==4.5.3
filetype==1.2.0
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.3
flask-restx==1.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
fonttools==4.60.1
fqdn==1.5.1
freezegun==1.5.5
frozenlist==1.8.0
gevent==25.5.1
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_resources==6.5.2
Incremental==24.11.0
inflection==0.5.1
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
kombu==5.6.2
laces==0.1.2
lark==1.3.1
Mako==1.3.10
Markdown==3.10.1
MarkupSafe==3.0.2
marshmallow==2.20.2
marshmallow-sqlalchemy==0.17.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
maxminddb==1.5.4
mistune==3.1.4
modelsearch==1.1.1
msgpack==1.1.2
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.3.0
notebook_shim==0.2.4
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
oracledb==3.4.1
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pillow==11.0.0
pillow_heif==1.1.1
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
psycopg2-binary==2.9.11
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
pymediainfo==7.0.1
PyMySQL==1.1.1
pynput==1.8.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pysubs2==1.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.13.0
python-geoacumen-city==2023.4.15
python-ipware==3.0.0
python-json-logger==4.0.0
pytz==2025.2
pyvidplayer2==0.9.28
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==5.0.1
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.0
s3transfer==0.10.4
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.1
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
telepath==0.3.1
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.0
uvicorn==0.38.0
vine==5.1.0
wagtail==7.2.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==2.2.3
whitenoise==6.6.0
Willow==1.12.0
WTForms==2.3.1
yarl==1.22.0
zope.event==5.1.1
zope.interface==7.2
